Refactor-1 wip
