+++ resources/js/Interfaces/Atribuicoes.ts
@@ -1,0 +1,6 @@
+import {Evento} from "@/Interfaces/Evento";
+
+export interface Atribuicoes {
+    nome_promotor: string;
+    atribuicoes: Evento[];
+}
