+++ resources/js/Interfaces/Espelho.ts
@@ -1,0 +1,10 @@
+import {Promotoria} from "@/Interfaces/Promotoria";
+
+export interface Espelho {
+    id: string;
+    periodo_inicio: string;
+    periodo_fim: string;
+    created_at: string;
+    updated_at: string;
+    promotorias?: Promotoria[];
+}
