+++ resources/js/Interfaces/Evento.ts
@@ -1,0 +1,9 @@
+export interface Evento {
+    uuid: string;
+    titulo: string;
+    tipo: string;
+    periodo_inicio: string;
+    periodo_fim: string;
+    promotor_titular_id: string;
+    promotor_designado_id: string;
+}
