+++ resources/js/Interfaces/HistoricoEventoServerSide.ts
@@ -1,0 +1,9 @@
+export interface HistoricoEventoServerSide {
+    id: string;
+    titulo: string;
+    tipo: string;
+    periodo_inicio: string;
+    periodo_fim: string;
+    historico_promotor_titular_id: string;
+    historico_promotor_designado_id: string;
+}
