+++ resources/js/Interfaces/HistoricoPromotoria.ts
@@ -1,0 +1,12 @@
+export interface HistoricoPromotoria {
+    id: string;
+    nome: string;
+    nome_grupo_promotorias: string;
+    municipio: string;
+    is_especializada: boolean;
+    historico_espelho_id: string;
+    historico_promotor_titular_id: string;
+    historico_id: string;
+    created_at: string;
+    updated_at: string;
+}
