+++ resources/js/Interfaces/HistoricoUrgenciaAtendimentoServeSide.ts
@@ -1,0 +1,6 @@
+export interface HistoricoUrgenciaAtendimentoServeSide {
+    id: string;
+    periodo_inicio: string;
+    periodo_fim: string;
+    historico_promotor_designado_id: string;
+}
