+++ resources/js/Interfaces/Promotor.ts
@@ -1,0 +1,7 @@
+export interface Promotor {
+    id: string;
+    nome: string;
+    is_substituto: boolean;
+    created_at: string;
+    updated_at: string;
+}
