+++ resources/js/Interfaces/Promotoria.ts
@@ -1,0 +1,10 @@
+export interface Promotoria {
+    id: string;
+    nome: string;
+    is_especializada: boolean;
+    espelho_id: string;
+    promotor_titular_id: string;
+    grupo_promotoria_id: string;
+    created_at: string;
+    updated_at: string;
+}
