+++ resources/js/Interfaces/UrgenciaAtendimentoClientSide.ts
@@ -1,0 +1,6 @@
+export interface UrgenciaAtendimentoClientSide {
+    uuid: string;
+    periodo_inicio: string;
+    periodo_fim: string;
+    promotor_designado_id: string;
+}
