+++ resources/js/Interfaces/UrgenciaAtendimentoServeSide.ts
@@ -1,0 +1,6 @@
+export interface UrgenciaAtendimentoServeSide {
+    id: string;
+    periodo_inicio: string;
+    periodo_fim: string;
+    promotor_designado_id: string;
+}
