+++ resources/js/types/index.d.ts
@@ -1,112 +1,8 @@
-export interface User {
-    id: string;
-    name: string;
-    email: string;
-    email_verified_at: string;
-}
-
-export interface Espelho {
-    id: string;
-    periodo_inicio: string;
-    periodo_fim: string;
-    created_at: string;
-    updated_at: string;
-    promotorias: Promotoria[];
-}
-
-export interface Promotor {
-    id: string;
-    nome: string;
-    is_substituto: boolean;
-    created_at: string;
-    updated_at: string;
-}
-
-export interface Promotoria {
-    id: string;
-    nome: string;
-    nome_grupo_promotorias: string;
-    municipio: string;
-    is_especializada: boolean;
-    espelho_id: string;
-    promotor_titular_id: string;
-    created_at: string;
-    updated_at: string;
-}
-
-export interface HistoricoPromotoria {
-    id: string;
-    nome: string;
-    nome_grupo_promotorias: string;
-    municipio: string;
-    is_especializada: boolean;
-    historico_espelho_id: string;
-    historico_promotor_titular_id: string;
-    created_at: string;
-    updated_at: string;
-}
-
-export interface EventoServerSide {
-    id: string;
-    titulo: string;
-    tipo: string;
-    periodo_inicio: string;
-    periodo_fim: string;
-    promotor_titular_id: string;
-    promotor_designado_id: string;
-}
-
-export interface HistoricoEventoServerSide {
-    id: string;
-    titulo: string;
-    tipo: string;
-    periodo_inicio: string;
-    periodo_fim: string;
-    historico_promotor_titular_id: string;
-    historico_promotor_designado_id: string;
-}
-
-export interface Evento {
-    uuid: string;
-    titulo: string;
-    tipo: string;
-    periodo_inicio: string;
-    periodo_fim: string;
-    promotor_titular_id: string;
-    promotor_designado_id: string;
-}
-
-export interface UrgenciaAtendimentoClientSide {
-    uuid: string;
-    periodo_inicio: string;
-    periodo_fim: string;
-    promotor_designado_id: string;
-}
-
-export interface UrgenciaAtendimentoServeSide {
-    id: string;
-    periodo_inicio: string;
-    periodo_fim: string;
-    promotor_designado_id: string;
-}
-
-export interface HistoricoUrgenciaAtendimentoServeSide {
-    id: string;
-    periodo_inicio: string;
-    periodo_fim: string;
-    historico_promotor_designado_id: string;
-}
-
-export interface GrupoPromotoria {
-    nome_grupo_promotorias: string;
-    promotorias: Promotoria[];
-    eventos: Evento[];
-}
-
-export interface Atribuicoes {
-    nome_promotor: string;
-    atribuicoes: Evento[];
-}
+import {User} from "@/Interfaces/User";
+import {Espelho} from "@/Interfaces/Espelho";
+import {Promotor} from "@/Interfaces/Promotor";
+import {Promotoria} from "@/Interfaces/Promotoria";
+import {Evento} from "@/Interfaces/Evento";
 
 export type PageProps<T extends Record<string, unknown> = Record<string, unknown>> = T & {
     auth: {
@@ -116,5 +12,4 @@ export type PageProps<T extends Record<string, unknown> = Record<string, unknown
     promotores: Promotor[];
     promotorias: Promotoria[];
     eventos: Evento[];
-    urgenciaAtendimentos: UrgenciaAtendimento[];
 };
